refactor(rest): extract FEA/results routes into routes/fea.py
Moves the four FEA-artefact/result routes (POST fea/artefacts, POST
fea/artefact, GET result-meta, GET fea/manifest) out of create_app's
closure into routes/fea.py, following the routes/__init__.py pattern.

Lifts the worker-advertised-extensions helper onto RestContext/deps.py
(a new worker_registry field plus deps.worker_advertised_exts) since
fea/manifest needs it; create_app keeps the old _worker_advertised_exts
name bound for the routes still in the closure.
